requirements.txt
streamlit
requests
sseclient-py
fastapi
uvicorn[standard]
